+++ src/RcppExports.cpp
@@ -54,8 +54,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // best_move_cpp
-NumericMatrix best_move_cpp(NumericMatrix ct, IntegerVector asg, int nstage, double lambda);
-RcppExport SEXP _stagedtrees_best_move_cpp(SEXP ctSEXP, SEXP asgSEXP, SEXP nstageSEXP, SEXP lambdaSEXP) {
+NumericMatrix best_move_cpp(NumericMatrix ct, IntegerVector asg, int nstage, double lambda, IntegerVector ctx);
+RcppExport SEXP _stagedtrees_best_move_cpp(SEXP ctSEXP, SEXP asgSEXP, SEXP nstageSEXP, SEXP lambdaSEXP, SEXP ctxSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -63,7 +63,8 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< IntegerVector >::type asg(asgSEXP);
     Rcpp::traits::input_parameter< int >::type nstage(nstageSEXP);
     Rcpp::traits::input_parameter< double >::type lambda(lambdaSEXP);
-    rcpp_result_gen = Rcpp::wrap(best_move_cpp(ct, asg, nstage, lambda));
+    Rcpp::traits::input_parameter< IntegerVector >::type ctx(ctxSEXP);
+    rcpp_result_gen = Rcpp::wrap(best_move_cpp(ct, asg, nstage, lambda, ctx));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -72,7 +73,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_stagedtrees_predict_lp_cpp", (DL_FUNC) &_stagedtrees_predict_lp_cpp, 5},
     {"_stagedtrees_path_lp_cpp", (DL_FUNC) &_stagedtrees_path_lp_cpp, 4},
     {"_stagedtrees_best_merge_cpp", (DL_FUNC) &_stagedtrees_best_merge_cpp, 4},
-    {"_stagedtrees_best_move_cpp", (DL_FUNC) &_stagedtrees_best_move_cpp, 4},
+    {"_stagedtrees_best_move_cpp", (DL_FUNC) &_stagedtrees_best_move_cpp, 5},
     {NULL, NULL, 0}
 };